% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualdubMain.R
\name{playAviMovieFile}
\alias{playAviMovieFile}
\title{Show AVI file in VLC media player}
\usage{
playAviMovieFile(avi)
}
\arguments{
\item{avi}{path to avi file}
}
\value{
shows avi file in VLC media player
}
\description{
Show AVI file in VLC media player
}
