% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualdubMain.R
\name{pngPictureFilesToAviMovieFile}
\alias{pngPictureFilesToAviMovieFile}
\title{Convert png files to avi movie using VirtualDub}
\usage{
pngPictureFilesToAviMovieFile(
  lastPngFile = dir(tdir, "^wtaqResult", full.names = TRUE)[1],
  tdir = kwb.utils::tempSubdirectory("wtaqResult"),
  aviFile = file.path(tdir, "out.avi"),
  virtualDub = system.file("extdata", "VirtualDub", "vdub.exe", package =
    "kwb.virtualdub"),
  show.avi = FALSE
)
}
\arguments{
\item{lastPngFile}{default: dir(tdir, "^wtaqResult", full.names=TRUE)[1]}

\item{tdir}{default: tempSubdirectory("wtaqResult")}

\item{aviFile}{default: file.path(tdir, "out.avi")}

\item{virtualDub}{path to virtualDub executable (default: 
system.file("extdata", "VirtualDub", "vdub.exe", package = "kwb.virtualdub"))}

\item{show.avi}{default: FALSE}
}
\value{
convert png files to avi movie
}
\description{
Convert png files to avi movie using VirtualDub
}
