% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualdubMain.R
\name{writeVirtualDubConfiguration}
\alias{writeVirtualDubConfiguration}
\title{Write configuration file for VirtualDub}
\usage{
writeVirtualDubConfiguration(outfile, nFrames)
}
\arguments{
\item{outfile}{outfile}

\item{nFrames}{nFrames}
}
\value{
Write VirtualDub configuration
}
\description{
Write configuration file for VirtualDub
}
